tools/novel_factory/pyproject.toml
[project]
name = "novel_factory"
version = "0.1.0"
description = "Offline CLI for serialized novel production"
authors = [{name="Exilens Tooling"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.21",
    "python-docx>=1.1",
    "pyperclip>=1.8",
]

[project.scripts]
novel-factory = "novel_factory.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
